requirements.txt
webob
gevent
requests
pyee
routes
nose
pyyaml
git+https://github.com/edgeware/mockito-python.git
behave
